# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stock Simulator API",
    "description": "Backend API for the stock market simulator",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "Game": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slot": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "PAUSED",
              "COMPLETED"
            ]
          },
          "step": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "completedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "totalPlayTime": {
            "type": "integer"
          },
          "character": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "gameId": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "uuid"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "name": {
                    "type": "string"
                  },
                  "balance": {
                    "type": "number"
                  },
                  "profession": {
                    "type": "string",
                    "enum": [
                      "DOCTOR",
                      "DEVELOPER",
                      "FINANCIER",
                      "FARMER",
                      "ENGINEER",
                      "STREET_CLEANER"
                    ]
                  },
                  "professionLevel": {
                    "type": "integer"
                  },
                  "salary": {
                    "type": "number"
                  },
                  "reputation": {
                    "type": "number"
                  },
                  "tradingLevel": {
                    "type": "integer"
                  },
                  "isNpc": {
                    "type": "boolean"
                  },
                  "dreamItemRefs": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "totalEarned": {
                    "type": "number"
                  },
                  "totalSpent": {
                    "type": "number"
                  },
                  "totalTrades": {
                    "type": "integer"
                  },
                  "successfulTrades": {
                    "type": "integer"
                  },
                  "createdAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "updatedAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "inventoryItems": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "characterId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "itemRef": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "purchasePrice": {
                          "type": "number"
                        },
                        "isInstallment": {
                          "type": "boolean"
                        },
                        "monthlyPayment": {
                          "type": [
                            "number",
                            "null"
                          ]
                        },
                        "installmentsTotal": {
                          "anyOf": [
                            {
                              "type": "integer"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "installmentsPaid": {
                          "type": "integer"
                        },
                        "special": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "isPaidOff": {
                          "type": "boolean"
                        },
                        "purchasedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "characterId",
                        "itemRef",
                        "name",
                        "purchasePrice",
                        "isInstallment",
                        "monthlyPayment",
                        "installmentsTotal",
                        "installmentsPaid",
                        "special",
                        "isPaidOff",
                        "purchasedAt"
                      ],
                      "additionalProperties": false
                    }
                  }
                },
                "required": [
                  "id",
                  "gameId",
                  "name",
                  "balance",
                  "profession",
                  "professionLevel",
                  "salary",
                  "reputation",
                  "tradingLevel",
                  "isNpc",
                  "dreamItemRefs",
                  "totalEarned",
                  "totalSpent",
                  "totalTrades",
                  "successfulTrades",
                  "createdAt",
                  "updatedAt"
                ],
                "additionalProperties": false
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "userId",
          "name",
          "slot",
          "status",
          "step",
          "createdAt",
          "updatedAt",
          "startedAt",
          "completedAt",
          "totalPlayTime",
          "character"
        ],
        "additionalProperties": false
      },
      "GameList": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid"
            },
            "userId": {
              "type": "string",
              "format": "uuid"
            },
            "name": {
              "type": "string"
            },
            "slot": {
              "type": "integer"
            },
            "status": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "INACTIVE",
                "PAUSED",
                "COMPLETED"
              ]
            },
            "step": {
              "type": "integer"
            },
            "createdAt": {
              "type": "string",
              "format": "date-time"
            },
            "updatedAt": {
              "type": "string",
              "format": "date-time"
            },
            "startedAt": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ]
            },
            "completedAt": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ]
            },
            "totalPlayTime": {
              "type": "integer"
            },
            "character": {
              "anyOf": [
                {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "gameId": {
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "uuid"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "name": {
                      "type": "string"
                    },
                    "balance": {
                      "type": "number"
                    },
                    "profession": {
                      "type": "string",
                      "enum": [
                        "DOCTOR",
                        "DEVELOPER",
                        "FINANCIER",
                        "FARMER",
                        "ENGINEER",
                        "STREET_CLEANER"
                      ]
                    },
                    "professionLevel": {
                      "type": "integer"
                    },
                    "salary": {
                      "type": "number"
                    },
                    "reputation": {
                      "type": "number"
                    },
                    "tradingLevel": {
                      "type": "integer"
                    },
                    "isNpc": {
                      "type": "boolean"
                    },
                    "dreamItemRefs": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "totalEarned": {
                      "type": "number"
                    },
                    "totalSpent": {
                      "type": "number"
                    },
                    "totalTrades": {
                      "type": "integer"
                    },
                    "successfulTrades": {
                      "type": "integer"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "inventoryItems": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "characterId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "itemRef": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "purchasePrice": {
                            "type": "number"
                          },
                          "isInstallment": {
                            "type": "boolean"
                          },
                          "monthlyPayment": {
                            "type": [
                              "number",
                              "null"
                            ]
                          },
                          "installmentsTotal": {
                            "anyOf": [
                              {
                                "type": "integer"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "installmentsPaid": {
                            "type": "integer"
                          },
                          "special": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "isPaidOff": {
                            "type": "boolean"
                          },
                          "purchasedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "characterId",
                          "itemRef",
                          "name",
                          "purchasePrice",
                          "isInstallment",
                          "monthlyPayment",
                          "installmentsTotal",
                          "installmentsPaid",
                          "special",
                          "isPaidOff",
                          "purchasedAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "id",
                    "gameId",
                    "name",
                    "balance",
                    "profession",
                    "professionLevel",
                    "salary",
                    "reputation",
                    "tradingLevel",
                    "isNpc",
                    "dreamItemRefs",
                    "totalEarned",
                    "totalSpent",
                    "totalTrades",
                    "successfulTrades",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                },
                {
                  "type": "null"
                }
              ]
            }
          },
          "required": [
            "id",
            "userId",
            "name",
            "slot",
            "status",
            "step",
            "createdAt",
            "updatedAt",
            "startedAt",
            "completedAt",
            "totalPlayTime",
            "character"
          ],
          "additionalProperties": false
        }
      },
      "CreateGameBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "slot": {
            "type": "integer",
            "minimum": 1,
            "maximum": 3
          },
          "profession": {
            "type": "string",
            "enum": [
              "DOCTOR",
              "DEVELOPER",
              "FINANCIER",
              "FARMER",
              "ENGINEER",
              "STREET_CLEANER"
            ]
          }
        },
        "required": [
          "name",
          "slot",
          "profession"
        ],
        "additionalProperties": false
      },
      "UpdateGameBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "PAUSED",
              "COMPLETED"
            ]
          }
        },
        "additionalProperties": false
      },
      "DeleteGameResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/saves": {
      "get": {
        "tags": [
          "saves"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameList"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "saves"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameBody"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Game"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/saves/{id}": {
      "get": {
        "tags": [
          "saves"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Game"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "saves"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGameBody"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Game"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "saves"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteGameResponse"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
}